third_party/WebKit/LayoutTests/http/tests/storage/async_cookies/document_getAll_set.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookes: cookieStore.getAll() sees cookieStore.set() cookie</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test(async testCase => {
  await cookieStore.set('cookie-name', 'cookie-value');
  const cookies = await cookieStore.getAll();

  assert_equals(cookies.length, 1);
  assert_equals(cookies[0].name, 'cookie-name');
  assert_equals(cookies[0].value, 'cookie-value');
}, 'cookieStore.getAll returns the cookie written by cookieStore.set');

</script>
